turning cookingTime from int to double

+++ MaghrebRecipes/ViewModel/AddRecipeViewModel.swift
@@ -12,12 +12,12 @@ class AddRecipeViewModel: ObservableObject {
     @Published var myRecipes: [RecipeModel] = []
     @Published var isSheetOn = false
     
-    func createRecipe(title : String, photo: String, description: String, allIngredients: [String], category : RecipeCategory, difficulty: RecipeDifficulty, averagePrice: RecipeAveragePrice, cookingTime: Int, timeToCook: TimeToCook) {
+    func createRecipe(title : String, photo: String, description: String, allIngredients: [String], category : RecipeCategory, difficulty: RecipeDifficulty, averagePrice: RecipeAveragePrice, cookingTime: Double, timeToCook: TimeToCook) {
         let recipe = RecipeModel(title: title, photo: photo, description: description,ingredients: allIngredients, recipCategory: category, recipeDifficulty: difficulty, recipeAveragePrice: averagePrice, valueTimeCooking: cookingTime, timeToCook: timeToCook)
         self.myRecipes.insert(recipe, at: 0)
     }
     
-    func addRecipeButton(title : String, photo: String, description: String, allIngredients: [String], category : RecipeCategory, difficulty: RecipeDifficulty, averagePrice: RecipeAveragePrice, cookingTime: Int, timeToCook: TimeToCook) {
+    func addRecipeButton(title : String, photo: String, description: String, allIngredients: [String], category : RecipeCategory, difficulty: RecipeDifficulty, averagePrice: RecipeAveragePrice, cookingTime: Double, timeToCook: TimeToCook) {
         createRecipe(title: title, photo: photo, description: description, allIngredients: allIngredients , category: category, difficulty: difficulty, averagePrice: averagePrice, cookingTime: cookingTime, timeToCook: timeToCook)
     }
